requirements-test.txt
-r requirements.txt
asynctest~=0.13
flake8~=4.0
flake8-docstrings~=1.6
mypy==0.930
pylint~=2.12
pylint-strict-informational==0.1
pytest~=6.2
pytest-cov~=2.12
pytest-homeassistant-custom-component>=0.5
